Include ppsi.mk in Makefile
so that ppsi is built directly by the wrc Makefile
(and not using a recursive make)

Fixes some format warning
(stdint.h now comes from include/std)

+++ boards/ertm14/board.c
@@ -2525,7 +2525,8 @@ static void mmc_show_version_info( const char *brdname, struct ertm14_mmc_state 
     pp_printf("MMC Build Info for %s:\n", brdname );
     pp_printf("  - Git build commit : %32s\n", st->info.git_sha );
     pp_printf("  - Git build tag    : %32s\n", st->info.git_tag );
-    pp_printf("  - Build date       : %lu (Unix)\n", le32_to_host( st->info.build_date ) );
+    pp_printf("  - Build date       : %u (Unix)\n",
+	      (unsigned)le32_to_host( st->info.build_date ) );
     pp_printf("  - Serial Number    : %32s\n",   st->info.board_serial_number );
 }
